backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
openai==1.3.0
bcrypt==4.1.1
PyJWT==2.8.0
numpy==1.26.2
scipy==1.11.4
pypandoc==1.12
python-docx==1.1.0
markdown==3.5.1
python-dotenv==1.0.0
reportlab==4.0.7
pillow>=10.0.0


